requirements.txt
fastapi==0.111.0
faiss-cpu==1.8.0.post1
pydantic==2.8.2
pydantic-settings==2.3.4
pandas==2.2.2
numpy==1.26.4
openai==1.35.13
langchain==0.2.7
requests==2.32.3
pytest==8.2.2
httpx==0.27.0
python-dotenv==1.0.1